Migrate request handling transaction into the database. Make a new table which contains all data about events which are going to be sent by event bus. Prepare migration rabbitmq to an external service

+++ src/WebStore.Infrastructure/Clients/ClientRepository.cs
@@ -1,6 +1,8 @@
 ﻿using FluentResults;
 using Microsoft.EntityFrameworkCore;
+using System.Text.Json;
 using WebStore.Domain;
+using WebStore.Infrastructure.RabbitMq.Events;
 
 namespace WebStore.Infrastructure.Clients
 {
@@ -8,33 +10,80 @@ namespace WebStore.Infrastructure.Clients
     {
         private readonly WebStoreContext _context = context;
 
-        public async Task AddClientAsync(Client client)
+        public async Task<Result> AddClientAsync(Client client)
         {
-            await _context.Clients.AddAsync(client);
-            await _context.SaveChangesAsync();
+            try
+            {
+                using var transaction = await _context.Database.BeginTransactionAsync();
+
+                await _context.Clients.AddAsync(client);
+                await _context.ClientEvents.AddAsync(CreateClientEvent(client, "client_created"));
+
+                await _context.SaveChangesAsync();
+
+                await transaction.CommitAsync();
+
+                return Result.Ok();
+            }
+            catch (Exception exception)
+            {
+                return Result.Fail(exception.Message);
+            }
         }
 
-        public async Task UpdateClientAsync(Client client)
+        public async Task<Result> UpdateClientAsync(Client client)
         {
             var entity = await _context.Clients.FindAsync(client.Id);
 
-            entity.Orders.Clear();
-            client.Orders.ForEach(entity.Orders.Add);
-            entity.Email = client.Email;
-            entity.PhoneNumber = client.PhoneNumber;
-            entity.Name = client.Name;
+            try
+            {
+                using var transaction = await _context.Database.BeginTransactionAsync();
 
-            await _context.SaveChangesAsync();
+                entity.Orders.Clear();
+                entity.Orders.AddRange(client.Orders);
+                entity.Email = client.Email;
+                entity.PhoneNumber = client.PhoneNumber;
+                entity.Name = client.Name;
+
+                await _context.ClientEvents.AddAsync(CreateClientEvent(client, "client_updated"));
+
+                await _context.SaveChangesAsync();
+
+                await transaction.CommitAsync();
+
+                return Result.Ok();
+            }
+            catch (Exception exception)
+            {
+                return Result.Fail(exception.Message);
+            }
+
         }
 
         public async Task<Result<Client>> DeleteClientAsync(Guid id)
         {
             var client = await _context.Clients.FindAsync(id);
 
-            _context.Clients.Remove(client);
-            await _context.SaveChangesAsync();
+            if (client is null)
+            {
+                return Result.Fail($"Client (ID:{id}) is not found");
+            }
 
-            return client;
+            try
+            {
+                using var transaction = await _context.Database.BeginTransactionAsync();
+                _context.Clients.Remove(client);
+                await _context.ClientEvents.AddAsync(CreateClientEvent(client, "client_deleted"));
+                await _context.SaveChangesAsync();
+
+                await transaction.CommitAsync();
+
+                return Result.Ok(client);
+            }
+            catch (Exception exception)
+            {
+                return Result.Fail(exception.Message);
+            }
         }
 
         public async Task<Result<Client>> GetClientAsync(Guid id)
@@ -51,7 +100,9 @@ namespace WebStore.Infrastructure.Clients
 
         public async Task<bool> IsPhoneNumberUniqueAsync(string phoneNumber) => !await _context.Clients.AnyAsync(x => x.PhoneNumber.Equals(phoneNumber));
         public async Task<bool> IsEmailUniqueAsync(string email) => !await _context.Clients.AnyAsync(x => x.Email.Equals(email));
-
         public async Task<Result<IEnumerable<Client>>> GetAllClientsAsync() => Result.Ok(await _context.Clients.ToListAsync() as IEnumerable<Client>);
+
+        private static ClientEvent CreateClientEvent(Client client, string queueName) => new(client.Id.ToString(), client.Name, client.PhoneNumber,
+            client.Email, "users/players/customers", queueName, JsonSerializer.Serialize(client.Orders));
     }
 }
